# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0",
      "title": "Free Mentors",
      "description": "This is a public Free mentors api",
      "contact": {}
    },
    "host": "free-mentors-app.herokuapp.com",
    "basePath": "/api/v1",
    "securityDefinitions": {},
    "schemes": [
      "http",
      "https"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
...
      "/auth/updateProfile": {
        "patch": {
          "description": "User can update the informations after authentification",
          "summary": "Update Profile",
          "tags": [
            "All Endpoints"
          ],
          "operationId": "updateprofile",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": false,
              "default": "Bearer {token}",
              "type": "string"
            },
            {
              "name": "Content-Type",
              "in": "header",
              "required": true,
              "type": "string",
              "description": ""
            },
            {
              "name": "occupation",
              "in": "formData",
              "required": true,
              "type": "string",
              "description": ""
            },
            {
                "name": "bio",
                "in": "formData",
                "required": true,
                "type": "string",
                "description": ""
              },
            {
              "name": "address",
              "in": "formData",
              "required": true,
              "type": "string",
              "description": ""
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          }
        }
      },
...
    },
...
  }